[update] refs #54 _componentsディレクトリ内整理
 開発体験改善のため
