+++ app/(routes)/(auth)/reset-password/page.tsx
@@ -1,7 +1,7 @@
 import { cookies } from 'next/headers'
 import { createServerComponentClient } from '@supabase/auth-helpers-nextjs'
 import { redirect } from 'next/navigation'
-import ResetPassword from '../../../_components/reset-password'
+import ResetPassword from '../../../_components/auth/reset-password'
 import type { Database } from '../../../../lib/database.types'
 import { dashboardPath } from '../../../_common/constants/path'
