+++ app/(routes)/(dashboard)/settings/password/page.tsx
@@ -1,7 +1,7 @@
 import { cookies } from 'next/headers'
 import { createServerComponentClient } from '@supabase/auth-helpers-nextjs'
 import { redirect } from 'next/navigation'
-import Password from '../../../../_components/password'
+import Password from '../../../../_components/settings/password'
 import type { Database } from '../../../../../lib/database.types'
 import { loginPath } from '../../../../_common/constants/path'
