+++ app/(routes)/(dashboard)/settings/profile/page.tsx
@@ -1,7 +1,7 @@
 import { cookies } from 'next/headers'
 import { createServerComponentClient } from '@supabase/auth-helpers-nextjs'
 import { redirect } from 'next/navigation'
-import Profile from '../../../../_components/profile'
+import Profile from '../../../../_components/settings/profile'
 
 import type { Database } from '../../../../../lib/database.types'
 import { loginPath } from '../../../../_common/constants/path'
