+++ app/_components/auth/signup.tsx
@@ -5,9 +5,9 @@ import { createClientComponentClient } from '@supabase/auth-helpers-nextjs'
 import { useRouter } from 'next/navigation'
 import { useForm, SubmitHandler } from 'react-hook-form'
 import { zodResolver } from '@hookform/resolvers/zod'
-import Loading from '../(routes)/loading'
+import Loading from '../../(routes)/loading'
 import * as z from 'zod'
-import type { Database } from '../../lib/database.types'
+import type { Database } from '../../../lib/database.types'
 type Schema = z.infer<typeof schema>
 import Image from 'next/image'
 import {
@@ -18,8 +18,8 @@ import {
   Typography,
   Input,
   Button,
-} from './common'
-import { callbackPath, loginPath } from '../_common/constants/path'
+} from '../common'
+import { callbackPath, loginPath } from '../../_common/constants/path'
 
 // 入力データの検証ルールを定義
 const schema = z.object({
