+++ app/_components/finance/acount-edit-form.tsx
@@ -1,11 +1,18 @@
 'use client'
-import { XMarkIcon } from '@heroicons/react/24/solid'
-import { Card, CardBody, Input, Button, Typography } from '../../common'
+import { TrashIcon, XMarkIcon } from '@heroicons/react/24/solid'
+import {
+  Card,
+  CardBody,
+  CardFooter,
+  Input,
+  Button,
+  Typography,
+} from '../common'
 import { Dispatch, SetStateAction, useState } from 'react'
 import { z } from 'zod'
-import Loading from '../../../(routes)/loading'
+import Loading from '../../(routes)/loading'
 import { createClientComponentClient } from '@supabase/auth-helpers-nextjs'
-import { Database } from '../../../../lib/database.types'
+import { Database } from '../../../lib/database.types'
 import { SubmitHandler, useForm } from 'react-hook-form'
 import { zodResolver } from '@hookform/resolvers/zod'
 import { useRouter } from 'next/navigation'
@@ -21,19 +28,32 @@ const schema = z.object({
   }),
 })
 
-export default function AcountAddForm({
-  open,
-  setOpen,
+type Acount = {
+  id: number
+  name: string
+  usage: string
+}
+
+export default function AcountEditForm({
+  show,
+  setShow,
+  acount,
   id,
 }: {
-  open: boolean
-  setOpen: Dispatch<SetStateAction<boolean>>
-  id: string
+  show: boolean
+  setShow: Dispatch<SetStateAction<boolean>>
+  acount: Acount
+  id: string | undefined
 }) {
   const router = useRouter()
   const supabase = createClientComponentClient<Database>()
   const [loading, setLoading] = useState(false)
   const [message, setMessage] = useState('')
+
+  const handleClose = () => {
+    setShow(false)
+    router.refresh()
+  }
 
   // 入力フォームの設定
   const {
@@ -43,12 +63,15 @@ export default function AcountAddForm({
     reset,
   } = useForm({
     // 初期値
-    defaultValues: { financeInstitution: '', usage: '' },
+    defaultValues: {
+      financeInstitution: acount.name,
+      usage: acount.usage,
+    },
     // 入力値の検証
     resolver: zodResolver(schema),
   })
 
-  // 送信
+  // 更新データを送信
   const onSubmit: SubmitHandler<Schema> = async (data) => {
     setLoading(true)
 
@@ -60,13 +83,14 @@ export default function AcountAddForm({
           .select('id')
           .eq('name', data.financeInstitution)
           .single()
-      // 登録済みの場合、メッセージ表示
-      if (FinancialInstitution !== null) {
-        setMessage('既に登録されています。')
-        return
-      }
-      if (FinancialInstitution === null) {
-        // 金融機関登録
+      // 登録済みの場合、金融機関更新
+      if (FinancialInstitution) {
+        const { error: errorUpdateFinancialInstitution } = await supabase
+          .from('FinancialInstitution')
+          .update({ usage: data.usage })
+          .eq('id', FinancialInstitution.id)
+        // 未登録の場合、金融機関登録
+      } else {
         const { error: errorInputFinancialInstitution } = await supabase
           .from('FinancialInstitution')
           .insert({
@@ -74,20 +98,18 @@ export default function AcountAddForm({
             name: data.financeInstitution,
             usage: data.usage,
           })
-          .eq('user_id', id)
+      }
 
-        // エラーチェック
-        if (errorInputFinancialInstitution) {
-          setMessage(
-            'エラーが発生しました。' + errorInputFinancialInstitution.message
-          )
-          return
-        }
+      // エラーチェック
+      if (errorFinancialInstitution) {
+        setMessage('エラーが発生しました。' + errorFinancialInstitution.message)
+        return
       }
 
       // 入力フォームクリア
       reset()
       setMessage('登録が完了しました。')
+      handleClose()
     } catch (error) {
       setMessage('エラーが発生しました。' + error)
       return
@@ -96,13 +118,40 @@ export default function AcountAddForm({
     }
   }
 
-  const handleClose = () => {
-    setOpen(false)
-    router.refresh()
+  // 金額削除
+  const handleDelete = async () => {
+    setLoading(true)
+
+    try {
+      // 金融機関削除
+      const { error: errorDeleteFinancialInstitution } = await supabase
+        .from('FinancialInstitution')
+        .delete()
+        .eq('user_id', id)
+        .eq('name', acount.name)
+
+      // エラーチェック
+      if (errorDeleteFinancialInstitution) {
+        setMessage(
+          'エラーが発生しました。' + errorDeleteFinancialInstitution.message
+        )
+        return
+      }
+
+      // 入力フォームクリア
+      reset()
+      setMessage('削除が完了しました。')
+      handleClose()
+    } catch (error) {
+      setMessage('エラーが発生しました。' + error)
+      return
+    } finally {
+      setLoading(false)
+    }
   }
 
   // 金額入力モーダルウィンドウを閉じる
-  if (!open) {
+  if (!show) {
     return null
   }
 
@@ -110,7 +159,7 @@ export default function AcountAddForm({
     <div className='fixed inset-0 z-40 overflow-y-auto w-screen h-screen bg-cyan-100 bg-opacity-25'>
       <div className='flex items-center justify-center min-h-screen pt-4 px-4 pb-20 text-center'>
         <Card>
-          <CardBody>
+          <CardBody className='pb-1'>
             <div className='text-end'>
               <Button
                 variant='text'
@@ -122,10 +171,10 @@ export default function AcountAddForm({
               </Button>
             </div>
             <Typography variant='h4' color='blue-gray'>
-              金融機関登録
+              金融機関変更
             </Typography>
             <Typography color='gray' className='mt-1 font-normal'>
-              Enter your financial institution to register
+              Enter your amounts to update or delete
             </Typography>
             <form
               onSubmit={handleSubmit(onSubmit)}
@@ -133,18 +182,19 @@ export default function AcountAddForm({
             >
               <div className='mb-4 flex flex-col gap-6'>
                 <Input
+                  type='text'
                   size='lg'
                   label='金融機関'
-                  type='text'
                   {...register('financeInstitution', { required: true })}
+                  readOnly
                 />
                 <div className='text-center text-sm text-red-500'>
                   {errors.financeInstitution?.message}
                 </div>
                 <Input
+                  type='text'
                   size='lg'
                   label='使い方'
-                  type='text'
                   {...register('usage', { required: true })}
                 />
                 <div className='text-center text-sm text-red-500'>
@@ -155,7 +205,7 @@ export default function AcountAddForm({
                 <Loading />
               ) : (
                 <Button type='submit' color='cyan' className='mt-6' fullWidth>
-                  Register
+                  Update
                 </Button>
               )}
               {message && (
@@ -165,6 +215,17 @@ export default function AcountAddForm({
               )}
             </form>
           </CardBody>
+          <CardFooter className='pt-0'>
+            <Button
+              onClick={handleDelete}
+              color='red'
+              className='flex items-center justify-center gap-3'
+              fullWidth
+            >
+              <TrashIcon fill='white' className='h-5 w-5'></TrashIcon>
+              Delete
+            </Button>
+          </CardFooter>
         </Card>
       </div>
     </div>
