+++ app/_components/finance/bar-chart.tsx
