+++ app/_components/finance/dashboard.tsx
@@ -1,9 +1,9 @@
 import FinanceBarChart from './bar-chart'
-import { Card, CardBody, CardFooter, Button, Typography } from '../../common'
+import { Card, CardBody, CardFooter, Button, Typography } from '../common'
 import AcountCard from './acount-card'
 import { createServerComponentClient } from '@supabase/auth-helpers-nextjs'
 import { cookies } from 'next/headers'
-import { Database } from '../../../../lib/database.types'
+import { Database } from '../../../lib/database.types'
 import AcountTable from './acount-table'
 import AcountAddButton from './acount-add-button'
 import AcountCardDefault from './acount-card-default'
