+++ app/_components/finance/finance-form.tsx
@@ -1,11 +1,11 @@
 'use client'
 import { XMarkIcon } from '@heroicons/react/24/solid'
-import { Card, CardBody, Input, Button, Typography } from '../../common'
+import { Card, CardBody, Input, Button, Typography } from '../common'
 import { Dispatch, SetStateAction, useState } from 'react'
 import { z } from 'zod'
-import Loading from '../../../(routes)/loading'
+import Loading from '../../(routes)/loading'
 import { createClientComponentClient } from '@supabase/auth-helpers-nextjs'
-import { Database } from '../../../../lib/database.types'
+import { Database } from '../../../lib/database.types'
 import { SubmitHandler, useForm } from 'react-hook-form'
 import { zodResolver } from '@hookform/resolvers/zod'
 import { useRouter } from 'next/navigation'
