+++ app/_components/settings/email.tsx
@@ -5,11 +5,11 @@ import { createClientComponentClient } from '@supabase/auth-helpers-nextjs'
 import { useRouter } from 'next/navigation'
 import { useForm, SubmitHandler } from 'react-hook-form'
 import { zodResolver } from '@hookform/resolvers/zod'
-import Loading from '../(routes)/loading'
+import Loading from '../../(routes)/loading'
 import * as z from 'zod'
-import type { Database } from '../../lib/database.types'
-import { Card, CardBody, Button } from './common'
-import { loginPath } from '../_common/constants/path'
+import type { Database } from '../../../lib/database.types'
+import { Card, CardBody, Button } from '../common'
+import { loginPath } from '../../_common/constants/path'
 type Schema = z.infer<typeof schema>
 
 // 入力データの検証ルールを定義
