+++ app/_components/total-assets/asset-actual-vs-target.tsx
@@ -1,5 +1,5 @@
-import { Database } from '../../../../lib/database.types'
-import { Button, Card, CardBody, CardFooter, Typography } from '../../common'
+import { Database } from '../../../lib/database.types'
+import { Button, Card, CardBody, CardFooter, Typography } from '../common'
 import { CurrencyYenIcon } from '@heroicons/react/24/solid'
 import GoalEditButton from './goal-edit-button'
 type GoalType = Database['public']['Tables']['Goal']['Row']
@@ -64,7 +64,7 @@ const AssetActualVsTarget = ({
       </CardBody>
       <CardFooter className='pt-0 pb-2'>
         <Typography className='text-base font-bold'>
-          目標金額 {goal?.amount ? goal?.amount?.toLocaleString(): 0}円
+          目標金額 {goal?.amount ? goal?.amount?.toLocaleString() : 0}円
         </Typography>
         <Typography className='text-2xl font-bold'>{goal?.goal}</Typography>
       </CardFooter>
