+++ app/_components/total-assets/asset-total-bar-chart.tsx
